+++ step16/step16/11866.cpp
@@ -1,0 +1,14 @@
+//#include <iostream>
+//#include <queue>
+//using namespace std;
+//
+//int main(void) {
+//	cin.tie(NULL);
+//	ios_base::sync_with_stdio(false);
+//
+//	int n, x;
+//	cin >> n;
+//	queue<int> qu;
+//
+//	return 0;
+//}
