+++ step16/step16/step16.vcxproj
@@ -129,6 +129,8 @@
   <ItemGroup>
     <ClCompile Include="10773.cpp" />
     <ClCompile Include="12789.cpp" />
+    <ClCompile Include="18258.cpp" />
+    <ClCompile Include="2164.cpp" />
     <ClCompile Include="28278.cpp" />
     <ClCompile Include="4949.cpp" />
     <ClCompile Include="9012.cpp" />
